+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 
 ## [Unreleased]
 
-- Fixed backspace not working in Termux
 - Implemented qwerty keyboard (swipe space bar to switch)
 - Implemented horizontal swiping of a key
+- Fixed backspace not working in Termux
 
 
 ## [v0.4.0] Symbols keyboard and fixes (2021-07-09)