+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Implemented qwerty keyboard (swipe space bar to switch)
 - Implemented horizontal swiping of a key
 - Fixed backspace not working in Termux
+- Improved main activity information text
 
 
 ## [v0.4.0] Symbols keyboard and fixes (2021-07-09)