+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Implemented horizontal swiping of a key
 - Fixed backspace not working in Termux
 - Improved main activity information text
+- Prevented keyboard from showing on main activity startup
 
 
 ## [v0.4.0] Symbols keyboard and fixes (2021-07-09)