+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [Unreleased]
 
+
+## [v0.5.0] Implemented qwerty (2021-07-12)
+
 - Implemented qwerty keyboard (swipe space bar to switch)
 - Implemented horizontal swiping of a key
 - Fixed backspace not working in Termux
@@ -66,7 +69,9 @@ the actual functionality has not been implemented yet.
 
 
 [Unreleased]:
-  https://github.com/stroke-input/stroke-input-android/compare/v0.4.0...HEAD
+  https://github.com/stroke-input/stroke-input-android/compare/v0.5.0...HEAD
+[v0.5.0]:
+  https://github.com/stroke-input/stroke-input-android/compare/v0.4.0...v0.5.0
 [v0.4.0]:
   https://github.com/stroke-input/stroke-input-android/compare/v0.3.0...v0.4.0
 [v0.3.0]: